% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_ply_add_theme.R
\name{ro_ply_add_theme}
\alias{ro_ply_add_theme}
\title{Add a Custom Theme to a Plotly Figure}
\usage{
ro_ply_add_theme(fig, theme)
}
\arguments{
\item{fig}{A Plotly figure object created by \code{plotly::plot_ly()} or similar.}

\item{theme}{A named list of layout parameters to apply as a theme (e.g., font, colors, margins).}
}
\value{
A Plotly figure object with the specified theme applied.
}
\description{
Applies a specified Plotly theme (as a list) to a Plotly figure object.
}
\examples{
\dontrun{
fig <- ro_ply_add_theme(fig, theme = theme_rivm)
}
}
\seealso{
Other plotly:
\code{\link[=ro_ply_cursor]{ro_ply_cursor()}},
\code{\link[=ro_ply_dim]{ro_ply_dim()}},
\code{\link[=ro_ply_disable_toolbar]{ro_ply_disable_toolbar()}},
\code{\link[=ro_ply_focus_linechart]{ro_ply_focus_linechart()}},
\code{\link[=ro_ply_keyboard_nav_barchart]{ro_ply_keyboard_nav_barchart()}},
\code{\link[=ro_ply_keyboard_nav_bargraph]{ro_ply_keyboard_nav_bargraph()}},
\code{\link[=ro_ply_keyboard_nav_legend]{ro_ply_keyboard_nav_legend()}},
\code{\link[=ro_ply_keyboard_nav_linechart]{ro_ply_keyboard_nav_linechart()}},
\code{\link[=ro_ply_keyboard_nav_trendline]{ro_ply_keyboard_nav_trendline()}},
\code{\link[=ro_ply_legend_theme]{ro_ply_legend_theme()}},
\code{\link[=ro_ply_theme]{ro_ply_theme()}},
\code{\link[=ro_ply_title]{ro_ply_title()}}
}
\concept{plotly}
