% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_ply_cursor.R
\name{ro_ply_cursor}
\alias{ro_ply_cursor}
\title{Set default cursor for a plotly figure}
\usage{
ro_ply_cursor(fig)
}
\arguments{
\item{fig}{A Plotly figure object created by \code{plotly::plot_ly()} or similar.}
}
\value{
A Plotly figure object with the cursor style set to "default" (arrow).
}
\description{
Changes the mouse cursor style for a Plotly figure from the default crosshair to the standard arrow cursor.
}
\examples{
\dontrun{
fig <- plotly::plot_ly(data = mtcars, x = ~mpg, y = ~hp, type = 'scatter', mode = 'markers')
fig_cursor <- ro_ply_cursor(fig)
fig_cursor
}
}
\seealso{
Other plotly:
\code{\link[=ro_ply_add_theme]{ro_ply_add_theme()}},
\code{\link[=ro_ply_dim]{ro_ply_dim()}},
\code{\link[=ro_ply_disable_toolbar]{ro_ply_disable_toolbar()}},
\code{\link[=ro_ply_focus_linechart]{ro_ply_focus_linechart()}},
\code{\link[=ro_ply_keyboard_nav_barchart]{ro_ply_keyboard_nav_barchart()}},
\code{\link[=ro_ply_keyboard_nav_bargraph]{ro_ply_keyboard_nav_bargraph()}},
\code{\link[=ro_ply_keyboard_nav_legend]{ro_ply_keyboard_nav_legend()}},
\code{\link[=ro_ply_keyboard_nav_linechart]{ro_ply_keyboard_nav_linechart()}},
\code{\link[=ro_ply_keyboard_nav_trendline]{ro_ply_keyboard_nav_trendline()}},
\code{\link[=ro_ply_legend_theme]{ro_ply_legend_theme()}},
\code{\link[=ro_ply_theme]{ro_ply_theme()}},
\code{\link[=ro_ply_title]{ro_ply_title()}}

Other ggplotly:
\code{\link[=ro_ply_disable_toolbar]{ro_ply_disable_toolbar()}},
\code{\link[=ro_ply_keyboard_nav_barchart]{ro_ply_keyboard_nav_barchart()}},
\code{\link[=ro_ply_keyboard_nav_bargraph]{ro_ply_keyboard_nav_bargraph()}},
\code{\link[=ro_ply_keyboard_nav_legend]{ro_ply_keyboard_nav_legend()}},
\code{\link[=ro_ply_keyboard_nav_linechart]{ro_ply_keyboard_nav_linechart()}},
\code{\link[=ro_ply_keyboard_nav_trendline]{ro_ply_keyboard_nav_trendline()}}
}
\concept{ggplotly}
\concept{plotly}
