% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_ply_dim.R
\name{ro_ply_dim}
\alias{ro_ply_dim}
\title{Set RIVM theme dimensions for a plotly Figure}
\usage{
ro_ply_dim(fig, width = 776, height = 400)
}
\arguments{
\item{fig}{A Plotly figure object created by \code{plotly::plot_ly()} or similar.}

\item{width}{Desired plot width in pixels. Default is 776.}

\item{height}{Desired plot height in pixels. Default is 400.}
}
\value{
A Plotly figure object with updated sizing policy and layout.
}
\description{
Applies custom width and height to a Plotly figure, ensuring the plot displays at the specified dimensions.
The function also sets the plot to autosize for responsive layouts.
}
\examples{
\dontrun{
fig <- plotly::plot_ly(data = mtcars, x = ~mpg, y = ~hp, type = 'scatter', mode = 'markers')
fig_dimmed <- ro_ply_dim(fig, width = 800, height = 500)
fig_dimmed
}
}
\seealso{
Other plotly:
\code{\link[=ro_ply_add_theme]{ro_ply_add_theme()}},
\code{\link[=ro_ply_cursor]{ro_ply_cursor()}},
\code{\link[=ro_ply_disable_toolbar]{ro_ply_disable_toolbar()}},
\code{\link[=ro_ply_focus_linechart]{ro_ply_focus_linechart()}},
\code{\link[=ro_ply_keyboard_nav_barchart]{ro_ply_keyboard_nav_barchart()}},
\code{\link[=ro_ply_keyboard_nav_bargraph]{ro_ply_keyboard_nav_bargraph()}},
\code{\link[=ro_ply_keyboard_nav_legend]{ro_ply_keyboard_nav_legend()}},
\code{\link[=ro_ply_keyboard_nav_linechart]{ro_ply_keyboard_nav_linechart()}},
\code{\link[=ro_ply_keyboard_nav_trendline]{ro_ply_keyboard_nav_trendline()}},
\code{\link[=ro_ply_legend_theme]{ro_ply_legend_theme()}},
\code{\link[=ro_ply_theme]{ro_ply_theme()}},
\code{\link[=ro_ply_title]{ro_ply_title()}}
}
\concept{plotly}
