% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_ply_legend_theme.R
\name{ro_ply_legend_theme}
\alias{ro_ply_legend_theme}
\title{RIVM Legend Styling for plotly figures}
\usage{
ro_ply_legend_theme(fig)
}
\arguments{
\item{fig}{A Plotly figure object created by \code{plotly::plot_ly()} or similar.}
}
\value{
The Plotly figure object with the custom legend styling applied.
}
\description{
Applies advanced custom styling to the legend of a Plotly figure.
This function adds a box around each individual legend item and changes legend symbols to rectangles.
The modifications are applied using JavaScript via \code{htmlwidgets::onRender()}.
}
\examples{
\dontrun{
fig <- plotly::plot_ly(data = mtcars, x = ~mpg, y = ~hp,
  color = ~factor(cyl), type = 'scatter', mode = 'markers')
fig_styled <- ro_ply_legend_theme(fig)
fig_styled
}
}
\seealso{
Other plotly:
\code{\link[=ro_ply_add_theme]{ro_ply_add_theme()}},
\code{\link[=ro_ply_cursor]{ro_ply_cursor()}},
\code{\link[=ro_ply_dim]{ro_ply_dim()}},
\code{\link[=ro_ply_disable_toolbar]{ro_ply_disable_toolbar()}},
\code{\link[=ro_ply_focus_linechart]{ro_ply_focus_linechart()}},
\code{\link[=ro_ply_keyboard_nav_barchart]{ro_ply_keyboard_nav_barchart()}},
\code{\link[=ro_ply_keyboard_nav_bargraph]{ro_ply_keyboard_nav_bargraph()}},
\code{\link[=ro_ply_keyboard_nav_legend]{ro_ply_keyboard_nav_legend()}},
\code{\link[=ro_ply_keyboard_nav_linechart]{ro_ply_keyboard_nav_linechart()}},
\code{\link[=ro_ply_keyboard_nav_trendline]{ro_ply_keyboard_nav_trendline()}},
\code{\link[=ro_ply_theme]{ro_ply_theme()}},
\code{\link[=ro_ply_title]{ro_ply_title()}}
}
\concept{plotly}
