% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_ply_title.R
\name{ro_ply_title}
\alias{ro_ply_title}
\title{Add Custom Titles and Axis Labels to a Plotly Figure}
\usage{
ro_ply_title(
  fig,
  y_data = NULL,
  y_label = NULL,
  x_label = NULL,
  title = NULL,
  base_family = "RijksoverheidSansWebText",
  canvas_width = 776,
  left_margin_base = 10,
  char_width = 7,
  title_size = 17,
  axis_title_size = 13,
  title_color = ro_color(color_name = "lintblauw"),
  thousand_seperator = ".",
  decimal_seperator = ","
)
}
\arguments{
\item{fig}{A plotly figure object created by \code{plotly::plot_ly()} to which the titles and labels will be added.}

\item{y_data}{Used to determine the longest y-axis label for margin calculation.}

\item{y_label}{Y-axis label to display.}

\item{x_label}{X-axis label to display.}

\item{title}{Main title to display at the top left.}

\item{base_family}{Font family to use.}

\item{canvas_width}{Width of the plotting area in pixels.}

\item{left_margin_base}{Base left margin in pixels.}

\item{char_width}{Estimated width per character in pixels.}

\item{title_size}{Font size for the main title.}

\item{axis_title_size}{Font size for axis titles.}

\item{title_color}{Color for all titles/labels.}

\item{thousand_seperator}{Separator for thousands in y-axis labels.}

\item{decimal_seperator}{Separator for decimals in y-axis labels.}
}
\value{
The plotly figure object with added (and styled) title and/or axis labels as annotations.
}
\description{
This function adds a (custom styled) main title, y-axis label, and/or x-axis label as annotations to a Plotly figure,
using RIVM theming and Dutch number formatting.
It automatically calculates left margin spacing based on the longest y-axis label.
}
\examples{
\dontrun{
# Title only:
ro_ply_title(fig,  y_data = data$'Number of Cases', title = "Fancy title")
# Y-axis label only:
ro_ply_title(fig, y_label = "Number of cases")
# X-axis label only:
ro_ply_title(fig, x_label = "Groups")
# Title, Y-axis label, and X-axis label in one go:
ro_ply_title(fig, y_data = data$'Number of Cases',
  y_label = "Number of cases", title = "Fancy title", x_label = "groups")
}
}
\seealso{
Other plotly:
\code{\link[=ro_ply_add_theme]{ro_ply_add_theme()}},
\code{\link[=ro_ply_cursor]{ro_ply_cursor()}},
\code{\link[=ro_ply_dim]{ro_ply_dim()}},
\code{\link[=ro_ply_disable_toolbar]{ro_ply_disable_toolbar()}},
\code{\link[=ro_ply_focus_linechart]{ro_ply_focus_linechart()}},
\code{\link[=ro_ply_keyboard_nav_barchart]{ro_ply_keyboard_nav_barchart()}},
\code{\link[=ro_ply_keyboard_nav_bargraph]{ro_ply_keyboard_nav_bargraph()}},
\code{\link[=ro_ply_keyboard_nav_legend]{ro_ply_keyboard_nav_legend()}},
\code{\link[=ro_ply_keyboard_nav_linechart]{ro_ply_keyboard_nav_linechart()}},
\code{\link[=ro_ply_keyboard_nav_trendline]{ro_ply_keyboard_nav_trendline()}},
\code{\link[=ro_ply_legend_theme]{ro_ply_legend_theme()}},
\code{\link[=ro_ply_theme]{ro_ply_theme()}}
}
\concept{plotly}
