% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROvis.plotly-package.R
\docType{package}
\name{ROvis.plotly-package}
\alias{ROvis.plotly}
\alias{ROvis.plotly-package}
\title{ROvis.plotly: Rijksoverheid Visualisatie - plotly}
\description{
ROvis.plotly is an R package that provides plotly visualisations in Rijksoverheid style.
}
\author{
\strong{Maintainer}: spin@rivm.nl \email{spin@rivm.nl}

Authors:
\itemize{
  \item spin@rivm.nl \email{spin@rivm.nl}
}

}
\keyword{internal}
